requirements.txt
reflex==0.8.20
google-genai
pillow